+++ ChangeLog.markdown
@@ -7,3 +7,4 @@ PHP_CodeCoverage 1.2.0
 ----------------------
 
 * PHP_CodeCoverage 1.2 is only supported on PHP 5.3.3 (or later) and PHP 5.4.0 (or later) is highly recommended.
+* Introduce the `@coversDefaultClass` annotation. It enables short @covers annotations when working with long class names or namespaces.